requirements.txt
alabaster==0.7.16
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
blinker==1.8.2
bounded-pool-executor==0.0.3
bs4==0.0.2
category-encoders==2.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
docutils==0.20.1
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
fonttools==4.51.0
graphviz==0.20.3
h11==0.14.0
idna==3.7
imagesize==1.4.1
imbalanced-learn==0.12.2
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.18.1
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaleido==0.2.1
kiwisolver==1.4.5
lightgbm==4.3.0
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
missingno==0.5.2
mysql-connector-python==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.59.1
numpy==1.23.5
numpydoc==1.6.0
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==1.5.2
parso==0.8.4
patsy==0.5.6
PDPbox==0.3.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
plotly-resampler==0.10.0
pluggy==1.5.0
pmdarima==2.0.4
pqdm==0.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycaret==3.1.0
pycparser==2.22
Pygments==2.18.0
pyod==1.1.3
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.1
schemdraw==0.15
scikit-base==0.5.2
scikit-learn==1.2.1
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.13.2
selenium==4.20.0
shap==0.41.0
six==1.16.0
sktime==0.21.1
slicer==0.0.7
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
statsmodels==0.14.2
sweetviz==2.3.1
tabulate==0.9.0
tbats==1.1.3
tenacity==8.3.0
threadpoolctl==3.5.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
tsdownsample==0.1.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
wurlitzer==3.1.0
xgboost==1.7.1
xxhash==3.4.1
yellowbrick==1.5
zipp==3.18.1
